server/requirements.txt
WTForms==2.2.1
Flask==1.0.2
Redis==2.10.6
configparser==3.5.0
psycopg2==2.7.4
postgres==2.2.1
SQLAlchemy==1.2.8
Flask-SQLAlchemy==2.3.2
Flask-Migrate
google-api-python-client
google-auth
requests
pytest
Flask-CORS
Flask-WTF==0.14.2
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.19.0